Update the Forms Dialog and Custom Dialog Forms
I tried to update the dialog form to have a message box display. This didn't work and all the different ways I tried displaying it failed. I feel I am missing something in the code or design. I posted a question to Stack Overflow about it. I tried copying the Dialog form for the custom dialog form that I was creating with a custom message window. This worked but had lots of conflict so I decided to place all the different forms into their own folders to possible resolve some of the naming issue. This work for some issue but I could not still fully copy the one form over it was still missing certain design elements in the actual design window but the code was copied but needed to be refactor to wire or map to the new design element names.