Dialog Form Update
I still can't get the message box to appear for either the button click event or for the form close event. I'm going to move on for a bit to the Custom Dialogs and see if I can come up with another solution. None of the research seems to yield anything and the code appears to run except for that function.