Built Custom Dialog for Window Close event
This dialog might solve my previous problem of not dialog appearing. I built a from with a panel that is docked at the top and from size is fixed. I disabled the controls on it and made two buttons one for yes and one for not and set their Dialog Result to be Yes and No.