Finished Custom Dialog and Refactored Dialog Code
I finished the Custom Dialog for the Confirm Close on the Dialog form. It still does not fire and I need to find the issue with that. I think it's a reference issue, I have a using statement but the code doesn't hit the Confirm Close on the Form when close. I'm going to move on an come back to it later.