Created New Show Form project with two forms.
I created a new project to test out why the show and show dialog where not working in my other forms? It has something to do with how my project is set up. I need to clean it up or rebuild them.
I was able to set up the regular show form and the show dialog that will not let you move between the two forms.