Created New Multiple Document Interface or MDI Text Editor Applicaiton
I built the Main Form.
I built the child form.
I added the Rich Text Editor to the main form and docked it to the main form.
I added a Menu Bar to the main form and a click event that would create a new child form or text editor document.