I updated the drop down windows and tool bar buttons to work correctly.
I added functionality to the windows dropdowns and toolbar buttons with event handlers to perform the actions they represent.
I need to refactor the code for the windows dropdown to work, it is currently based on the visual basic code and doesn't show a list of open forms like it should.